requirements.txt
# Core Framework & Database
Django==5.2.7
mysqlclient>=2.2.1
django-environ>=0.13,<0.14

# API & Frontend Communication (New)
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
requests>=2.31.0

# Authentication & OAuth (From earlier)
django-allauth>=0.57.0

PyJWT
cryptography
